requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
twilio==6.45.2
urllib3==1.25.10
